DDCNL-8759: Moved private methods to helper
